Rough draft of the etch-a-sketch w/ html css & js For the html document I added a container div that will hold the grid.
The javascript provided captures that grid container element and gives
it a default 16x16 grid using a for loop. This for loop also provides
the hover effect in order for the grid squares to change color when you
mouse over them. The next for loop will replace the old grid with a new
one based on user input provided through a button.

The css styesheet provides the flexbox which makes the grid possible and
sets some boundaries.